- removed duplicated headerHandlerStub implementations
