+++ epochStart/metachain/systemSCs_test.go
@@ -307,7 +307,7 @@ func TestSystemSCProcessor_UpdateStakingV2ShouldWork(t *testing.T) {
 	doStake(t, s.systemVM, s.userAccountsDB, owner1, big.NewInt(1000), blsKeys[0], blsKeys[1])
 	doStake(t, s.systemVM, s.userAccountsDB, owner2, big.NewInt(1000), blsKeys[2], blsKeys[3])
 
-	args.EpochNotifier.CheckEpoch(&mock.HeaderHandlerStub{
+	args.EpochNotifier.CheckEpoch(&testscommon.HeaderHandlerStub{
 		EpochField: 1000000,
 	})
 
@@ -372,7 +372,7 @@ func TestSystemSCProcessor_UpdateStakingV2MoreKeysShouldWork(t *testing.T) {
 	}
 	sw.Stop("do stake")
 
-	args.EpochNotifier.CheckEpoch(&mock.HeaderHandlerStub{
+	args.EpochNotifier.CheckEpoch(&testscommon.HeaderHandlerStub{
 		EpochField: 1000000,
 	})
 
@@ -1216,7 +1216,7 @@ func TestSystemSCProcessor_ProcessSystemSmartContractMaxNodesStakedFromQueueOwne
 		make([]byte, 0),
 	)
 
-	args.EpochNotifier.CheckEpoch(&mock.HeaderHandlerStub{
+	args.EpochNotifier.CheckEpoch(&testscommon.HeaderHandlerStub{
 		EpochField: 10,
 	})
 	validatorInfos := make(map[uint32][]*state.ValidatorInfo)
